project4/requirements.txt
Flask==2.0.3
Flask_RESTful==0.3.9
numpy==1.24.3
Pillow==9.5.0
Requests==2.31.0
streamlit==1.12.0
torch==1.10.0
torchvision==0.11.1
